requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
continuous-threading==2.0.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
playsound==1.2.2
psutil==5.8.0
pyserial==3.5
pyrocket==0.2.9
requests==2.25.1
simpleaudio==1.0.4
urllib3==1.26.6
waitress==2.0.0
Werkzeug==2.0.1
